gym.make() with specstack
